% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_summary.R
\name{better_summary}
\alias{better_summary}
\title{Better summary}
\usage{
better_summary(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
An improved version of \code{summary} from base R.
}
\details{
Base R \code{summary} returns a table object that is not friendly for
analysis. \code{better_summary} looks to improve on this by following
Tidyverse principles.
}
\examples{
data("iris")
better_summary(iris)
}
